iiod: Move Avahi support to its own dns-sd.c source file
Avoid having a huge #ifdef block in iiod.c by moving this code to a
dns-sd.c source file, that will be compiled conditionally.
